# server/swagger.yaml
swagger: 2.0.0
info:
  title: Way Farer
  description: WayFarer is a public bus transportation booking service.
  version: 0.1.9
host: "way-farer-andela.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT_Token_Admin:
    type: apiKey
    in: header
    name: token
  JWT_Token_User:
    type: apiKey
    in: header
    name: token
paths:
...
  /trips:
    post:
      security:
        - JWT_Token_Admin: []
      tags:
        - "Trip operations"
      summary: Create a new trips.
      operationId: createTrip
      parameters:
        - name: "token"
          in: "header"
          required: true
          schema:
            $ref: "#/definitions/TripCreate"
      responses:
        '200':
          description: OK
    get:
      security:
        - JWT_Token_User: []
      tags:
        - "Trip operations"
      summary: Get all trips.
      operationId: getTrip
      parameters:
        - in: "header"
          name: "token"
          description: "New trip object created"
          required: true
          schema:
            $ref: "#/definitions/TripList"
      responses:
        '200':
          description: OK
...
definitions:
...
  TripCreate:
    type: object
    properties:
      token:
        type: string
      seating_capacity:
        type: string
      origin:
        type: string
      destination:
        type: string
      trip_date:
        type: string
      fare:
        type: string
  TripList:
    type: object
    properties:
      id:
        type: integer
      seating_capacity:
        type: string
      origin:
         type: string
      destination:
         type: string
      trip_date:
         type: string
      fare:
         type: string
...